requirements.txt
streamlit==1.29.0
chromadb==0.4.18
groq==0.4.1
python-dotenv==1.0.0
langchain==0.1.0
sentence-transformers==2.2.2
rank-bm25==0.2.2
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
plotly==5.18.0